+++ src/Eoss.Backend.Application/CloudSense/Ptz/PtzAppService.cs
@@ -149,7 +149,7 @@ namespace Eoss.Backend.CloudSense.Ptz
         }
 
         public async Task<PtzStatusInDegreeDto> AbsoluteMoveWithDegreeAsync(string deviceId, string profileToken, 
-            float panInDegree, float tiltInDegree, int zoomInLevel, float panSpeed = 1, float tiltSpeed = 1, float zoomSpeed = 1)
+            float panInDegree, float tiltInDegree, float zoomInLevel, float panSpeed = 1, float tiltSpeed = 1, float zoomSpeed = 1)
         {
             try
             {
@@ -175,7 +175,7 @@ namespace Eoss.Backend.CloudSense.Ptz
         }
 
         public async Task<PtzStatusInDegreeDto> RelativeMoveWithDegreeAsync(string deviceId, string profileToken, 
-            float panInDegree, float tiltInDegree, int zoomInLevel, float panSpeed = 1, float tiltSpeed = 1, float zoomSpeed = 1)
+            float panInDegree, float tiltInDegree, float zoomInLevel, float panSpeed = 1, float tiltSpeed = 1, float zoomSpeed = 1)
         {
             try
             {
